Nlp/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
redis
openai
